Shutdown refactoring
