+++ src/main.cpp
@@ -20,6 +20,7 @@
 #include "main.h"
 #include "modules/Modules.h"
 #include "sleep.h"
+#include "shutdown.h"
 #include "target_specific.h"
 #include "debug/i2cScan.h"
 #include "debug/axpDebug.h"
@@ -435,45 +436,8 @@ void setup()
     setCPUFast(false); // 80MHz is fine for our slow peripherals
 }
 
-
 uint32_t rebootAtMsec; // If not zero we will reboot at this time (used to reboot shortly after the update completes)
 uint32_t shutdownAtMsec; // If not zero we will shutdown at this time (used to shutdown from python or mobile client)
-
-void powerCommandsCheck()
-{
-    if (rebootAtMsec && millis() > rebootAtMsec) {
-#ifndef NO_ESP32
-        DEBUG_MSG("Rebooting for update\n");
-        ESP.restart();
-#else
-        DEBUG_MSG("FIXME implement reboot for this platform");
-#endif
-    }
-
-#if NRF52_SERIES
-    if (shutdownAtMsec) {
-        screen->startShutdownScreen();
-        playBeep();
-        ledOff(PIN_LED1);
-        ledOff(PIN_LED2);
-    }
-#endif
-
-    if (shutdownAtMsec && millis() > shutdownAtMsec) {
-        DEBUG_MSG("Shutting down from admin command\n");
-#ifdef TBEAM_V10
-        if (axp192_found == true) {
-            setLed(false);
-            power->shutdown();
-        }
-#elif NRF52_SERIES
-        playShutdownMelody();
-        power->shutdown();
-#else
-        DEBUG_MSG("FIXME implement shutdown for this platform");
-#endif
-    }
-}
 
 // If a thread does something that might need for it to be rescheduled ASAP it can set this flag
 // This will supress the current delay and instead try to run ASAP.
