+++ src/shutdown.h
@@ -1,0 +1,41 @@
+#include "configuration.h"
+#include "graphics/Screen.h"
+#include "power.h"
+#include "buzz.h"
+#include "main.h"
+
+void powerCommandsCheck()
+{
+    if (rebootAtMsec && millis() > rebootAtMsec) {
+#ifndef NO_ESP32
+        DEBUG_MSG("Rebooting for update\n");
+        ESP.restart();
+#else
+        DEBUG_MSG("FIXME implement reboot for this platform");
+#endif
+    }
+
+#if NRF52_SERIES
+    if (shutdownAtMsec) {
+        screen->startShutdownScreen();
+        playBeep();
+        ledOff(PIN_LED1);
+        ledOff(PIN_LED2);
+    }
+#endif
+
+    if (shutdownAtMsec && millis() > shutdownAtMsec) {
+        DEBUG_MSG("Shutting down from admin command\n");
+#ifdef TBEAM_V10
+        if (axp192_found == true) {
+            setLed(false);
+            power->shutdown();
+        }
+#elif NRF52_SERIES
+        playShutdownMelody();
+        power->shutdown();
+#else
+        DEBUG_MSG("FIXME implement shutdown for this platform");
+#endif
+    }
+}
